now we can transform the image
and there is something we can do to remove the red corners. just need to
move some pixel inside the pic so the coners are out, it is a good thing
to do. and need to combine all the functionblity to the main dialog, so we
will have the same pic or video in the config window... i suggest using
pic or just one frame in the config... oh fk... such a long comment. i
wish my d**k is such long.....